Sebaszm/lifetime caching stubs

+++ Source/com/IStringIterator.h
@@ -118,7 +118,7 @@ namespace RPC {
         }
         virtual bool Next(string& result) override
         {
-            uint32_t length = _container.size();
+            uint32_t length = static_cast<uint32_t>(_container.size());
 
             if (_index <= length) {
                 _index++;
